workers/requirements.txt
# Celery and dependencies
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# For legacy check compatibility
gitpython==3.1.41

# Development/Linting tools (optional, for development)
black==24.2.0
flake8==7.0.0
flake8-bugbear==24.2.6
isort==5.13.2
mypy==1.8.0
pylint==3.0.3
pytest==7.4.4
